Development: Move data-export DTOs out of core into their producing modules
Relocate the GDPR data-export DTOs from core/dto/export to the modules that
produce them, so the data lives inside its own (toggleable) module boundary
instead of in always-on core:
- atlas/dto/export: CompetencyProgressExportDTO, UserCompetencyProgressExportDTO, LearnerProfileExportDTO, UserLearnerProfileExportDTO
- iris/dto/export: IrisChatSessionExportDTO, IrisMessageExportDTO

The now-empty core/dto/export package is removed. DataExportResource stays at
api/core/data-exports (user-facing endpoint, client path unchanged). admin/core
consumers reach the DTOs via the modules' externally-allowed dto packages, so
the atlas/iris api-boundary arch tests still pass. The grade/score DTO cluster
(CourseGradeScoreDTO, ExamGradeScoreDTO, ...) is intentionally left for a
separate PR (it requires editing JPQL constructor-projection strings).
