+++ src/main/java/de/tum/cit/aet/artemis/atlas/api/CompetencyProgressApi.java
@@ -12,11 +12,11 @@ import de.tum.cit.aet.artemis.account.domain.User;
 import de.tum.cit.aet.artemis.atlas.config.AtlasEnabled;
 import de.tum.cit.aet.artemis.atlas.domain.LearningObject;
 import de.tum.cit.aet.artemis.atlas.domain.competency.CourseCompetency;
+import de.tum.cit.aet.artemis.atlas.dto.export.CompetencyProgressExportDTO;
+import de.tum.cit.aet.artemis.atlas.dto.export.UserCompetencyProgressExportDTO;
 import de.tum.cit.aet.artemis.atlas.repository.CompetencyProgressRepository;
 import de.tum.cit.aet.artemis.atlas.repository.CompetencyRepository;
 import de.tum.cit.aet.artemis.atlas.service.competency.CompetencyProgressService;
-import de.tum.cit.aet.artemis.core.dto.export.CompetencyProgressExportDTO;
-import de.tum.cit.aet.artemis.core.dto.export.UserCompetencyProgressExportDTO;
 import de.tum.cit.aet.artemis.exercise.domain.participation.Participant;
 
 @Controller
