+++ src/main/java/de/tum/cit/aet/artemis/atlas/api/LearnerProfileApi.java
@@ -10,12 +10,12 @@ import org.springframework.stereotype.Controller;
 import de.tum.cit.aet.artemis.account.domain.User;
 import de.tum.cit.aet.artemis.atlas.config.AtlasEnabled;
 import de.tum.cit.aet.artemis.atlas.domain.profile.LearnerProfile;
+import de.tum.cit.aet.artemis.atlas.dto.export.LearnerProfileExportDTO;
+import de.tum.cit.aet.artemis.atlas.dto.export.UserLearnerProfileExportDTO;
 import de.tum.cit.aet.artemis.atlas.repository.CourseLearnerProfileRepository;
 import de.tum.cit.aet.artemis.atlas.repository.LearnerProfileRepository;
 import de.tum.cit.aet.artemis.atlas.service.profile.CourseLearnerProfileService;
 import de.tum.cit.aet.artemis.atlas.service.profile.LearnerProfileService;
-import de.tum.cit.aet.artemis.core.dto.export.LearnerProfileExportDTO;
-import de.tum.cit.aet.artemis.core.dto.export.UserLearnerProfileExportDTO;
 import de.tum.cit.aet.artemis.course.domain.Course;
 
 @Controller
